Add exceptPlatforms to AMP links and tracking parameter reference tests
