+++ privacy-config/privacy-config-impl/src/test/java/com/duckduckgo/privacy/config/impl/referencetests/trackingparameters/TrackingParameterReferenceTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 DuckDuckGo
+ * Copyright (c) 2023 DuckDuckGo
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,16 +14,18 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.privacy.config.impl.features.amplinks
+package com.duckduckgo.privacy.config.impl.referencetests.trackingparameters
 
 import com.duckduckgo.app.FileUtilities
 import com.duckduckgo.app.privacy.db.UserAllowListRepository
 import com.duckduckgo.feature.toggles.api.FeatureToggle
-import com.duckduckgo.privacy.config.api.AmpLinkException
-import com.duckduckgo.privacy.config.api.AmpLinks
 import com.duckduckgo.privacy.config.api.PrivacyFeatureName
+import com.duckduckgo.privacy.config.api.TrackingParameterException
+import com.duckduckgo.privacy.config.api.TrackingParameters
 import com.duckduckgo.privacy.config.api.UnprotectedTemporary
-import com.duckduckgo.privacy.config.store.features.amplinks.AmpLinksRepository
+import com.duckduckgo.privacy.config.impl.features.trackingparameters.RealTrackingParameters
+import com.duckduckgo.privacy.config.impl.features.trackingparameters.TrackingParametersFeature
+import com.duckduckgo.privacy.config.store.features.trackingparameters.TrackingParametersRepository
 import com.squareup.moshi.JsonAdapter
 import com.squareup.moshi.Moshi
 import java.util.concurrent.CopyOnWriteArrayList
@@ -38,21 +40,21 @@ import org.mockito.kotlin.whenever
 import org.robolectric.ParameterizedRobolectricTestRunner
 
 @RunWith(ParameterizedRobolectricTestRunner::class)
-class AmpKeywordReferenceTest(private val testCase: TestCase) {
+class TrackingParameterReferenceTest(private val testCase: TestCase) {
 
-    lateinit var testee: AmpLinks
+    lateinit var testee: TrackingParameters
 
-    private val mockRepository: AmpLinksRepository = mock()
+    private val mockRepository: TrackingParametersRepository = mock()
     private val mockUnprotectedTemporary: UnprotectedTemporary = mock()
     private val mockFeatureToggle: FeatureToggle = mock()
     private val mockUserAllowListRepository: UserAllowListRepository = mock()
 
     @Before
     fun setup() {
-        mockAmpLinks()
-        testee = RealAmpLinks(mockRepository, mockFeatureToggle, mockUnprotectedTemporary, mockUserAllowListRepository)
+        mockTrackingParameters()
+        testee = RealTrackingParameters(mockRepository, mockFeatureToggle, mockUnprotectedTemporary, mockUserAllowListRepository)
         whenever(mockUnprotectedTemporary.isAnException(any())).thenReturn(false)
-        whenever(mockFeatureToggle.isFeatureEnabled(PrivacyFeatureName.AmpLinksFeatureName.value, true)).thenReturn(true)
+        whenever(mockFeatureToggle.isFeatureEnabled(PrivacyFeatureName.TrackingParametersFeatureName.value, true)).thenReturn(true)
     }
 
     companion object {
@@ -64,53 +66,53 @@ class AmpKeywordReferenceTest(private val testCase: TestCase) {
         fun testData(): List<TestCase> {
             val test = adapter.fromJson(
                 FileUtilities.loadText(
-                    AmpKeywordReferenceTest::class.java.classLoader!!,
-                    "reference_tests/amplinks/tests.json",
+                    TrackingParameterReferenceTest::class.java.classLoader!!,
+                    "reference_tests/trackingparameters/tests.json",
                 ),
             )
-            return test?.ampKeywords?.tests ?: emptyList()
+            return test?.trackingParameters?.tests?.filterNot { it.exceptPlatforms.contains("android-browser") } ?: emptyList()
         }
     }
 
     @Test
     fun whenReferenceTestRunsItReturnsTheExpectedResult() {
-        val extractedUrl = testee.extractCanonicalFromAmpLink(testCase.ampURL)
-        if (extractedUrl != null) {
-            assertTrue(testCase.expectAmpDetected)
+        val cleanedUrl = testee.cleanTrackingParameters(initiatingUrl = testCase.initiatorURL, url = testCase.testURL)
+        if (cleanedUrl != null) {
+            assertEquals(testCase.expectURL, cleanedUrl)
         } else {
-            assertFalse(testCase.expectAmpDetected)
+            assertEquals(testCase.expectURL, testCase.testURL)
         }
     }
 
-    private fun mockAmpLinks() {
-        val jsonAdapter: JsonAdapter<AmpLinksFeature> = moshi.adapter(AmpLinksFeature::class.java)
-        val exceptions = CopyOnWriteArrayList<AmpLinkException>()
-        val ampLinkKeywords = CopyOnWriteArrayList<String>()
+    private fun mockTrackingParameters() {
+        val jsonAdapter: JsonAdapter<TrackingParametersFeature> = moshi.adapter(TrackingParametersFeature::class.java)
+        val exceptions = CopyOnWriteArrayList<TrackingParameterException>()
+        val trackingParameters = CopyOnWriteArrayList<Regex>()
         val jsonObject: JSONObject = FileUtilities.getJsonObjectFromFile(
-            AmpKeywordReferenceTest::class.java.classLoader!!,
-            "reference_tests/amplinks/config_reference.json",
+            TrackingParameterReferenceTest::class.java.classLoader!!,
+            "reference_tests/trackingparameters/config_reference.json",
         )
 
         val features: JSONObject = jsonObject.getJSONObject("features")
 
-        features.keys().forEach {
-            val ampLinksFeature: AmpLinksFeature? = jsonAdapter.fromJson(features.get(it).toString())
-            exceptions.addAll(ampLinksFeature!!.exceptions)
-            ampLinkKeywords.addAll(ampLinksFeature.settings.keywords)
+        features.keys().forEach { key ->
+            val trackingParametersFeature: TrackingParametersFeature? = jsonAdapter.fromJson(features.get(key).toString())
+            exceptions.addAll(trackingParametersFeature!!.exceptions)
+            trackingParameters.addAll(trackingParametersFeature.settings.parameters.map { it.toRegex() })
         }
         whenever(mockRepository.exceptions).thenReturn(exceptions)
-        whenever(mockRepository.ampLinkFormats).thenReturn(CopyOnWriteArrayList())
-        whenever(mockRepository.ampKeywords).thenReturn(ampLinkKeywords)
+        whenever(mockRepository.parameters).thenReturn(trackingParameters)
     }
 
     data class TestCase(
         val name: String,
-        val ampURL: String,
-        val expectAmpDetected: Boolean,
+        val testURL: String,
+        val initiatorURL: String,
+        val expectURL: String,
         val exceptPlatforms: List<String>,
     )
 
-    data class AmpKeywordTest(
+    data class TrackingParameterTest(
         val name: String,
         val desc: String,
         val referenceConfig: String,
@@ -118,6 +120,6 @@ class AmpKeywordReferenceTest(private val testCase: TestCase) {
     )
 
     data class ReferenceTest(
-        val ampKeywords: AmpKeywordTest,
+        val trackingParameters: TrackingParameterTest,
     )
 }
